# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_AlertOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AlertOut"
            },
            "title": "Items",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "items",
          "meta"
        ],
        "title": "Page[AlertOut]",
        "type": "object"
      },
      "Page_AuditLogOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditLogOut"
            },
            "title": "Items",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "items",
          "meta"
        ],
        "title": "Page[AuditLogOut]",
        "type": "object"
      },
      "Page_UserOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UserOut"
            },
            "title": "Items",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "items",
          "meta"
        ],
        "title": "Page[UserOut]",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "Fraud-scoring backend API. See /docs for interactive OpenAPI docs and README.md for architecture, known limitations, and what is mocked vs. real in the current implementation.",
    "title": "FraudGuard ML API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/audit": {
      "get": {
        "operationId": "list_audit_v1_admin_audit_get",
        "parameters": [
          {
            "description": "Filter by exact action name",
            "in": "query",
            "name": "action",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by exact action name",
              "title": "Action"
            }
          },
          {
            "description": "Filter by acting user id",
            "in": "query",
            "name": "actor_user_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by acting user id",
              "title": "Actor User Id"
            }
          },
          {
            "description": "Sort field, prefix with '-' for descending. One of: created_at, -created_at",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "-created_at",
              "description": "Sort field, prefix with '-' for descending. One of: created_at, -created_at",
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "description": "1-indexed page number",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "1-indexed page number",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Items per page",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "description": "Items per page",
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AuditLogOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Audit",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/v1/admin/users": {
      "get": {
        "operationId": "list_users_v1_admin_users_get",
        "parameters": [
          {
            "description": "Filter by role. One of: ['analyst', 'investigator', 'admin', 'manager']",
            "in": "query",
            "name": "role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by role. One of: ['analyst', 'investigator', 'admin', 'manager']",
              "title": "Role"
            }
          },
          {
            "description": "Sort field, prefix with '-' for descending. One of: created_at, -created_at, email, -email",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "-created_at",
              "description": "Sort field, prefix with '-' for descending. One of: created_at, -created_at, email, -email",
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "description": "1-indexed page number",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "1-indexed page number",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Items per page",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "description": "Items per page",
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_UserOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Users",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/v1/fraud/alerts": {
      "get": {
        "operationId": "alerts_feed_v1_fraud_alerts_get",
        "parameters": [
          {
            "description": "Filter by decision. One of: ['allow', 'challenge', 'block']",
            "in": "query",
            "name": "decision",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by decision. One of: ['allow', 'challenge', 'block']",
              "title": "Decision"
            }
          },
          {
            "description": "Sort field, prefix with '-' for descending. One of: created_at, -created_at, risk_score, -risk_score",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "-created_at",
              "description": "Sort field, prefix with '-' for descending. One of: created_at, -created_at, risk_score, -risk_score",
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "description": "1-indexed page number",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "1-indexed page number",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Items per page",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "description": "Items per page",
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AlertOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Alerts Feed",
        "tags": [
          "fraud"
        ]
      }
    },
    "/v1/fraud/behavior": {
      "post": {
        "operationId": "track_behavior_v1_fraud_behavior_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BehaviorEventIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Track Behavior",
        "tags": [
          "fraud"
        ]
      }
    },
    "/v1/fraud/device": {
      "post": {
        "operationId": "register_device_v1_fraud_device_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Register Device",
        "tags": [
          "fraud"
        ]
      }
    },
    "/v1/fraud/feedback": {
      "post": {
        "operationId": "submit_feedback_v1_fraud_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Submit Feedback",
        "tags": [
          "fraud"
        ]
      }
    },
    "/v1/fraud/graph": {
      "get": {
        "description": "Builds a user/device/IP relationship graph for investigator tooling\nfrom one page of recent transactions.\n\nAPI-05 FIX: previously used a hardcoded `.limit(GRAPH_MAX_TRANSACTIONS)`\ninstead of this API's shared pagination contract. Now accepts the same\n`page`/`page_size` query parameters as every other list endpoint\n(`/fraud/alerts`, `/admin/audit`, `/admin/users`) via `PageParams`, and\na client pages through transactions (most-recent-first) instead of\nalways receiving one fixed top-N slice.\n\nThis remains a synchronous, request-time computation, intentionally\nscoped to one bounded page for fast interactive use. For the *full*\ngraph across up to GRAPH_JOB_MAX_TRANSACTIONS transactions, use\n`POST /graph/jobs` instead (SCALE-02 background-job path) \u2014 see that\nendpoint's docstring.",
        "operationId": "fraud_graph_v1_fraud_graph_get",
        "parameters": [
          {
            "description": "1-indexed page number",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "1-indexed page number",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Items per page",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "description": "Items per page",
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Fraud Graph",
        "tags": [
          "fraud"
        ]
      }
    },
    "/v1/fraud/graph/jobs": {
      "post": {
        "description": "SCALE-02 FIX: computing the *full* fraud-ring graph (as opposed to one\npaginated page \u2014 see GET /graph above) can mean scanning up to\nGRAPH_JOB_MAX_TRANSACTIONS rows and building a networkx graph from all\nof them, which is exactly the kind of work that should not block a\nrequest/response cycle or an API worker. This kicks the computation off\nvia FastAPI's `BackgroundTasks` (chosen over Celery+Redis: this project\nis not otherwise running a message broker, and BackgroundTasks is\nexplicitly the right-sized tool for \"run this after the response, in\nthe same process\" without introducing new infrastructure) and returns\nimmediately with a job id; poll `GET /graph/jobs/{job_id}` for status\nand, once done, the result.",
        "operationId": "create_graph_job_v1_fraud_graph_jobs_post",
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Graph Job",
        "tags": [
          "fraud"
        ]
      }
    },
...
  }
}